complete pubsub communication confirmation
