+++ src/services/jobQueue/jobQueueName.ts
@@ -1,0 +1,4 @@
+export enum JobQueueName {
+    GD_ORG_REVIEW_RENEWAL_CRONJOB = 'gdOrgReviewRenewalCronjob',
+    GD_ORG_REVIEW_SCRAPER_JOB = 'gdOrgReviewScraperJob'
+}
